+++ de.metas.swat/de.metas.swat.base/src/main/java/de/metas/adempiere/form/terminal/swing/SwingTerminalScrollPane.java
@@ -30,6 +30,7 @@ import java.awt.Component;
 import java.awt.Dimension;
 
 import javax.swing.JScrollPane;
+import javax.swing.JViewport;
 import javax.swing.border.Border;
 
 import org.compiere.swing.CScrollPane;